+++ client/app/login/login.tsx
@@ -1,0 +1,7 @@
+
+
+export const login = () => {
+  console.log("Login Button Clicked");
+}
+
+export default login;
